+++ src/graphql/resolvers/customer.ts
@@ -5,14 +5,16 @@ import {
   CustomerInput,
   CustomerUpdateInput,
 } from "../../schemas/customer.zod.js";
-import { Customer } from "../../models/customer.js";
+import { Customer, ICustomer } from "../../models/customer.js";
 import {
   NotFoundError,
   ValidationError,
   DatabaseError,
   isValidationError,
+  AuthenticationError,
 } from "../../utils/errors.js";
 import { GraphQLError } from "graphql";
+import { GraphQLContext } from "../context.js";
 
 export const customerResolvers = {
   Query: {
@@ -44,7 +46,12 @@ export const customerResolvers = {
     },
   },
   Mutation: {
-    createCustomer: async (_: unknown, { input }: { input: CustomerInput }) => {
+    createCustomer: async (
+      _: unknown,
+      { input }: { input: CustomerInput },
+      context: GraphQLContext
+    ) => {
+      if (!context.customer) throw new AuthenticationError();
       try {
         const validatedData = validateCustomer(input);
         const customer = new Customer(validatedData);
@@ -58,16 +65,26 @@ export const customerResolvers = {
     },
     updateCustomer: async (
       _: unknown,
-      { id, input }: { id: string; input: CustomerUpdateInput }
+      { id, input }: { id: string; input: CustomerUpdateInput },
+      context: GraphQLContext
     ) => {
+      if (!context.customer) throw new AuthenticationError();
       try {
         const validatedData = validateCustomerUpdate(input);
-        const customer = await Customer.findByIdAndUpdate(id, validatedData, {
-          new: true,
-        });
+        const customer = await Customer.findOneAndUpdate(
+          { _id: id },
+          { $set: validatedData },
+          {
+            new: true,
+            runValidators: true,
+          }
+        );
         if (!customer) {
           throw new NotFoundError("Customer");
         }
+        // Manually check and update accountSetupComplete
+        customer.accountSetupComplete = customer.isAccountSetupComplete();
+        await customer.save();
         return customer;
       } catch (error) {
         if (error instanceof NotFoundError) {
@@ -79,7 +96,12 @@ export const customerResolvers = {
         throw new DatabaseError("Failed to update customer");
       }
     },
-    deleteCustomer: async (_: unknown, { id }: { id: string }) => {
+    deleteCustomer: async (
+      _: unknown,
+      { id }: { id: string },
+      context: GraphQLContext
+    ) => {
+      if (!context.customer) throw new AuthenticationError();
       try {
         const customer = await Customer.findByIdAndDelete(id);
         if (!customer) {
@@ -93,5 +115,9 @@ export const customerResolvers = {
         throw new DatabaseError("Failed to delete customer");
       }
     },
+  },
+  Customer: {
+    accountSetupComplete: (parent: ICustomer) =>
+      parent.accountSetupComplete ?? false,
   },
 };
